TADA-Compute/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.13.2
pymongo==4.3.3
pandas==1.3.5
requests==2.28.2
pycoingecko==3.1.0
azure-storage-blob==12.14.1
azure-storage-queue==12.5.0
logging-azure==1.0.2
html-to-json==2.0.0
DateTime==4.9
python-time==0.3.0
numpy==1.24.1
PyPDF2==2.11.1
opencensus-ext-azure==1.1.8
tqdm==4.64.1